fix(dahsboard): made dashboard start when opening
